Add ProxyCAFile in ProxyConfig struct
User will know where this CA comes from now.
